requirements.txt
# Core dependencies for Python 3.11
opencv-python>=4.8.1
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
matplotlib>=3.7.0

# GUI and utilities
tqdm>=4.66.0

# Optional dependencies
xformers>=0.0.22; sys_platform != "darwin"  # Memory optimization (not available on macOS ARM)

# Development and testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0 
